refactor(tokenizer): Extract `parser.lexer.src_handler` module
